Add PHPUnit testing and refine code structure
This commit introduces PHPUnit for project testing and refines the structure of Env.php for parseValue method efficiency. It also updates the .gitignore and composer.json files with new entries for PHPUnit and vendor directory respectively. A new EnvTest.php file has been added to set up the testing framework.